# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MBKM API” Manajemen User & Role",
    "version": "1.1.0",
    "description": "Admin mengelola akun dosen (Tim MBKM, GPM, Penguji Semhas, DPL, dll - satu dosen boleh punya beberapa role) dan akun mahasiswa langsung dibuat oleh admin. Klik Authorize lalu isi token dari /auth/login untuk mencoba endpoint yang butuh login."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Mitra": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "nama": {
            "type": "string",
            "example": "PT. Mencari Cinta Sejati"
          },
          "alamat": {
            "type": "string",
            "nullable": true,
            "example": "Jalan Mawar No. 5"
          },
          "email": {
            "type": "string",
            "nullable": true,
            "example": "MencariCinta@gmail.com"
          },
          "deskripsi": {
            "type": "string",
            "nullable": true,
            "example": "Lorem ipsum dolor sit amet"
          },
          "foto_url": {
            "type": "string",
            "nullable": true,
            "example": "https://xxx.supabase.co/storage/v1/object/public/mitra-foto/foto.jpg"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateMitraRequest": {
        "type": "object",
        "required": [
          "nama"
        ],
        "properties": {
          "nama": {
            "type": "string",
            "example": "PT. Mencari Cinta Sejati"
          },
          "alamat": {
            "type": "string",
            "example": "Jalan Mawar No. 5"
          },
          "email": {
            "type": "string",
            "example": "MencariCinta@gmail.com"
          },
          "deskripsi": {
            "type": "string",
            "example": "Perusahaan bergerak di bidang teknologi"
          },
          "foto": {
            "type": "string",
            "format": "binary",
            "description": "File gambar (JPG/PNG/WebP/GIF, max 5 MB)"
          }
        }
      },
      "UpdateMitraRequest": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string",
            "example": "PT. Nama Baru"
          },
          "alamat": {
            "type": "string",
            "example": "Jalan Mawar No. 10"
          },
          "email": {
            "type": "string",
            "example": "baru@gmail.com"
          },
          "deskripsi": {
            "type": "string",
            "example": "Deskripsi baru"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "foto": {
            "type": "string",
            "format": "binary",
            "description": "File gambar baru (opsional, foto lama otomatis terhapus)"
          }
        }
      },
      "CreateMitraJsonRequest": {
        "type": "object",
        "required": [
          "nama"
        ],
        "properties": {
          "nama": {
            "type": "string",
            "example": "PT. Mencari Cinta Sejati"
          },
          "alamat": {
            "type": "string",
            "example": "Jalan Mawar No. 5"
          },
          "email": {
            "type": "string",
            "example": "MencariCinta@gmail.com"
          },
          "deskripsi": {
            "type": "string",
            "example": "Perusahaan teknologi"
          }
        }
      },
...
      "CreatePaketKonversiRequest": {
        "type": "object",
        "required": [
          "nama"
        ],
        "properties": {
          "nama": {
            "type": "string",
            "description": "Nama plotting / role / skill MBKM",
            "example": "UI/UX DESIGN"
          },
          "deskripsi": {
            "type": "string",
            "example": "Paket konversi mata kuliah untuk role UI/UX Designer"
          },
          "mata_kuliah_ids": {
            "type": "array",
            "description": "Daftar ID mata kuliah (UUID) yang dipilih dari dropdown",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "example": [
              "a0000000-0000-0000-0000-000000000001",
              "a0000000-0000-0000-0000-000000000002"
            ]
          }
        }
      },
      "UpdatePaketKonversiRequest": {
        "type": "object",
        "description": "Minimal satu field. Jika mata_kuliah_ids dikirim, seluruh daftar mata kuliah pada paket akan diganti.",
        "properties": {
          "nama": {
            "type": "string",
            "example": "UI/UX DESIGN & RESEARCH"
          },
          "deskripsi": {
            "type": "string"
          },
          "mata_kuliah_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/mitra": {
      "get": {
        "tags": [
          "Mitra"
        ],
        "summary": "List semua mitra magang (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Cari berdasarkan nama/email/alamat"
          },
          {
            "name": "is_active",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "rows": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Mitra"
                          }
                        },
                        "total": {
                          "type": "integer",
                          "example": 25
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi"
          },
          "403": {
            "description": "Bukan admin"
          }
        }
      },
      "post": {
        "tags": [
          "Mitra"
        ],
        "summary": "Tambah mitra magang baru (admin)",
        "description": "Kirim sebagai multipart/form-data. Field foto opsional.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateMitraRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMitraJsonRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mitra berhasil ditambahkan",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Mitra"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal"
          },
          "401": {
            "description": "Tidak terautentikasi"
          },
          "403": {
            "description": "Bukan admin"
          }
        }
      }
    },
    "/mitra/{id}": {
      "get": {
        "tags": [
          "Mitra"
        ],
        "summary": "Detail mitra magang (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Mitra"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Mitra tidak ditemukan"
          }
        }
      },
      "put": {
        "tags": [
          "Mitra"
        ],
        "summary": "Edit data mitra magang (admin)",
        "description": "Kirim sebagai multipart/form-data. Field foto opsional — jika diisi, foto lama otomatis diganti.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMitraRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMitraJsonRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Mitra berhasil diperbarui",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Mitra"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validasi gagal"
          },
          "404": {
            "description": "Mitra tidak ditemukan"
          }
        }
      },
      "delete": {
        "tags": [
          "Mitra"
        ],
        "summary": "Hapus mitra magang (admin)",
        "description": "Foto di Supabase Storage juga ikut terhapus.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Mitra berhasil dihapus"
          },
          "404": {
            "description": "Mitra tidak ditemukan"
          }
        }
      }
    },
...
    "/paket-konversi": {
      "get": {
        "tags": [
          "Paket Konversi"
        ],
        "summary": "Daftar paket konversi mata kuliah (plot role/skill MBKM)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10,
              "maximum": 100
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Cari berdasarkan nama plotting atau deskripsi paket"
          }
        ],
        "responses": {
          "400": {
            "description": "Validasi gagal"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin atau tim MBKM"
          },
          "200": {
            "description": "OK: {rows,total}"
          }
        }
      },
      "post": {
        "tags": [
          "Paket Konversi"
        ],
        "summary": "Buat paket konversi baru (plot role/skill MBKM)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaketKonversiRequest"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Validasi gagal atau ID mata kuliah tidak ditemukan"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin atau tim MBKM"
          },
          "201": {
            "description": "Paket konversi berhasil dibuat"
          }
        }
      }
    },
    "/paket-konversi/{id}": {
      "get": {
        "tags": [
          "Paket Konversi"
        ],
        "summary": "Detail paket konversi beserta daftar mata kuliah terpilih",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID paket konversi"
          }
        ],
        "responses": {
          "400": {
            "description": "Validasi ID gagal"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin atau tim MBKM"
          },
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Paket konversi tidak ditemukan"
          }
        }
      },
      "put": {
        "tags": [
          "Paket Konversi"
        ],
        "summary": "Edit paket konversi & sinkronisasi item mata kuliah",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID paket konversi"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePaketKonversiRequest"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Validasi gagal atau ID mata kuliah tidak ditemukan"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin atau tim MBKM"
          },
          "200": {
            "description": "Paket konversi berhasil diperbarui"
          },
          "404": {
            "description": "Paket konversi tidak ditemukan"
          }
        }
      },
      "delete": {
        "tags": [
          "Paket Konversi"
        ],
        "summary": "Hapus paket konversi",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "ID paket konversi"
          }
        ],
        "responses": {
          "400": {
            "description": "Validasi ID gagal"
          },
          "401": {
            "description": "Belum login"
          },
          "403": {
            "description": "Bukan admin atau tim MBKM"
          },
          "200": {
            "description": "Paket konversi berhasil dihapus"
          },
          "404": {
            "description": "Paket konversi tidak ditemukan"
          }
        }
      }
    }
  }
}